Move Card into its own component folder
