+++ src/components/Search.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import axios from 'axios';
-import Card from './Card';
+import Card from './Card/';
 import Filters from './Filters';
 
 class Search extends Component {
